requirements.txt
discord.py
loguru
PyYAML
aiosqlite
pyfiglet
pystyle
paramiko